+++ index.html
@@ -7,7 +7,7 @@
   <link rel="stylesheet" href="css/style.css">
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <script src="js/functions.js"></script>
-  <script src="js/main.js"></script>
+  <script type="module" src="main.js"></script>
   <title>Кекстаграм</title>
 </head>
