+++ js/util.js
@@ -1,0 +1,7 @@
+// функция для генерации случайного числа в заданном диапазоне
+export function getRandomInt(a, b) {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
